Commitment creation is complete, just need data store
