+++ lib/goal_model.dart
@@ -1,9 +1,10 @@
 class GoalModel {
   String name;
   String description;
+  DateTime targetDate;
   int targetAmount;
   int currentAmount;
 
-  GoalModel(
-      this.name, this.description, this.targetAmount, this.currentAmount) {}
+  GoalModel(this.name, this.description, this.targetDate, this.targetAmount,
+      this.currentAmount) {}
 }
